OpenCorePkg/Library/OcRtcLib: fix possible loss of data
OpenCorePkg\Library\OcRtcLib\AppleRtcRam.c(151): error C2220: the following warning is treated as an error
OpenCorePkg\Library\OcRtcLib\AppleRtcRam.c(151): warning C4244: 'function': conversion from 'UINTN' to 'UINT8', possible loss of data

+++ Library/OcRtcLib/AppleRtcRam.c
@@ -148,7 +148,7 @@ AppleRtcRamReadData (
       return Status;
     }
 
-    Status = SyncRtcRead (Address, Buffer);
+    Status = SyncRtcRead ((UINT8) Address, Buffer);
     if (EFI_ERROR (Status)) {
       return Status;
     }
